+++ front/src/components/map/MapInfo.tsx
@@ -1,0 +1,5 @@
+import { Box } from "@chakra-ui/react"
+
+export const MapInfo = ({ eventName }) => {
+  return <Box>{eventName}</Box>
+}
